refactor(ios): consolidate compact-and-apply via MutationCompactor; retire MutationQueue.enqueue
Promotes the eager-compaction-plus-stage pattern into a single static helper —
`MutationCompactor.compactAndApply(_:in:)` — so `ItemStore`, `ListStore`, and
`ShareIngestor` all flow through one source of truth. Each call site previously
carried its own private `applyCompacted` helper, three near-identical copies.

The production `MutationQueue.enqueue` is gone: it had a single caller left
(ShareIngestor) and its built-in `save()` was the original Wave B atomicity
hazard — the optimistic SwiftData write and the queue-entry insert had
already been pulled into the stores' transactions. ShareIngestor now stages
its compaction-aware deltas the same way and the drain's tail `context.save()`
persists them in one transaction.

Test surface kept intact via a `BrettTests/TestSupport` extension that
preserves the legacy `MutationQueue.enqueue` shape — the queue-lifecycle
tests (FIFO, dependsOn ordering, fail/retry, complete, resetInFlight) still
read clearly without rewriting every case to construct entries by hand.
