+++ apps/ios/Brett/Stores/ItemStore.swift
@@ -305,7 +305,7 @@ final class ItemStore: Clearable {
             method: .post,
             payload: JSONCodec.encode(payload)
         )
-        applyCompacted(entityType: "item", entityId: item.id, incoming: entry)
+        MutationCompactor.compactAndApply(entry, in: context)
     }
 
     private func enqueueUpdate(
@@ -328,7 +328,7 @@ final class ItemStore: Clearable {
             baseUpdatedAt: item._baseUpdatedAt,
             beforeSnapshot: JSONCodec.encode(beforeSnapshot)
         )
-        applyCompacted(entityType: "item", entityId: item.id, incoming: entry)
+        MutationCompactor.compactAndApply(entry, in: context)
     }
 
     private func enqueueDelete(_ item: Item, beforeSnapshot: [String: Any]) {
@@ -342,70 +342,9 @@ final class ItemStore: Clearable {
             baseUpdatedAt: item._baseUpdatedAt,
             beforeSnapshot: JSONCodec.encode(beforeSnapshot)
         )
-        applyCompacted(entityType: "item", entityId: item.id, incoming: entry)
+        MutationCompactor.compactAndApply(entry, in: context)
     }
 
-    /// Run eager compaction for the given entity, then apply the resulting
-    /// deltas to `context` WITHOUT calling save. `MutationCompactor.compact`
-    /// is a pure value function — it returns ids to delete, an existing entry
-    /// it mutated in place, and/or a new entry to insert.
-    ///
-    /// This used to live inside `MutationQueue.enqueue`, but that helper
-    /// commits a save mid-method, which split the optimistic SwiftData write
-    /// from the queue-entry insert across two transactions and broke the
-    /// atomic-rollback guarantee. By doing the compaction here — between the
-    /// optimistic write and the outer `saver.save()` — the whole thing
-    /// commits or rolls back together.
-    ///
-    /// Without compaction, 10 rapid edits to the same field produce 10 queue
-    /// rows that the push engine then has to round-trip individually. With
-    /// it, they collapse to a single UPDATE.
-    private func applyCompacted(
-        entityType: String,
-        entityId: String,
-        incoming: MutationQueueEntry
-    ) {
-        let pending = fetchPendingMutations(entityType: entityType, entityId: entityId)
-        let result = MutationCompactor.compact(pending: pending, incoming: incoming)
-
-        for id in result.toDelete {
-            if let entry = fetchMutationEntry(id: id) {
-                context.delete(entry)
-            }
-        }
-        // `result.toUpdate`, if non-nil, is an already-persisted entry the
-        // compactor mutated in place. SwiftData re-persists the dirty fields
-        // on the next save, so we don't need to do anything explicit here.
-        if let toInsert = result.toInsert {
-            context.insert(toInsert)
-        }
-    }
-
-    private func fetchPendingMutations(
-        entityType: String,
-        entityId: String
-    ) -> [MutationQueueEntry] {
-        let pendingRaw = MutationStatus.pending.rawValue
-        var descriptor = FetchDescriptor<MutationQueueEntry>(
-            predicate: #Predicate {
-                $0.entityType == entityType
-                    && $0.entityId == entityId
-                    && $0.status == pendingRaw
-            },
-            sortBy: [SortDescriptor(\.createdAt, order: .forward)]
-        )
-        descriptor.includePendingChanges = true
-        return (try? context.fetch(descriptor)) ?? []
-    }
-
-    private func fetchMutationEntry(id: String) -> MutationQueueEntry? {
-        var descriptor = FetchDescriptor<MutationQueueEntry>(
-            predicate: #Predicate { $0.id == id }
-        )
-        descriptor.fetchLimit = 1
-        descriptor.includePendingChanges = true
-        return try? context.fetch(descriptor).first
-    }
 }
 
 // MARK: - Shared JSON helper
