+++ apps/ios/Brett/Sync/MutationQueue.swift
@@ -4,9 +4,10 @@ import SwiftData
 /// Durable FIFO queue of pending writes, backed by `MutationQueueEntry` rows
 /// in SwiftData. Mirrors `apps/mobile/src/sync/mutation-queue.ts`.
 ///
-/// - Stores call `enqueue(...)` after applying a local optimistic write.
-///   Enqueue runs eager compaction so redundant mutations are collapsed
-///   before they ever reach the push engine.
+/// - Stores stage new entries via `MutationCompactor.compactAndApply(_:in:)`
+///   alongside their optimistic SwiftData write, then commit both in a single
+///   `context.save()`. Compaction collapses redundant mutations before they
+///   ever reach the push engine.
 /// - The push engine (W2-B) drains pending entries via `pendingEntries(limit:)`,
 ///   flips them to in-flight via `markInFlight(ids:)`, then calls
 ///   `complete(id:)` or `fail(id:...)` depending on server response.
@@ -22,78 +23,14 @@ final class MutationQueue: MutationQueueProtocol {
     /// Matches spec §2.4 ("retryCount >= 10 → dead").
     static let maxRetries = 10
 
-    private let context: ModelContext
+    /// Visible at module level (not `private`) so test-only helpers can stage
+    /// a `MutationCompactor.compactAndApply(_:in:)` + save against the same
+    /// context for queue-lifecycle setup. Production callers go through the
+    /// store's own context — they don't read this.
+    let context: ModelContext
 
     init(context: ModelContext) {
         self.context = context
-    }
-
-    // MARK: - Enqueue
-
-    /// Insert a new pending mutation and run eager compaction against any
-    /// existing pending rows for the same entity.
-    ///
-    /// Returns the persisted entry (possibly compacted into an older one).
-    /// The caller does not need to save the context; we save before returning.
-    @discardableResult
-    func enqueue(
-        entityType: String,
-        entityId: String,
-        action: MutationAction,
-        endpoint: String,
-        method: MutationMethod,
-        payload: String,
-        changedFields: String? = nil,
-        previousValues: String? = nil,
-        baseUpdatedAt: String? = nil,
-        beforeSnapshot: String? = nil,
-        dependsOn: String? = nil,
-        batchId: String? = nil,
-        idempotencyKey: String? = nil,
-        now: Date = Date()
-    ) -> MutationQueueEntry? {
-        let incoming = MutationQueueEntry(
-            idempotencyKey: idempotencyKey,
-            entityType: entityType,
-            entityId: entityId,
-            action: action,
-            endpoint: endpoint,
-            method: method,
-            payload: payload,
-            changedFields: changedFields,
-            previousValues: previousValues,
-            baseUpdatedAt: baseUpdatedAt,
-            beforeSnapshot: beforeSnapshot,
-            dependsOn: dependsOn,
-            batchId: batchId,
-            createdAt: now
-        )
-
-        let existing = pendingForEntity(entityType: entityType, entityId: entityId)
-        let result = MutationCompactor.compact(pending: existing, incoming: incoming)
-
-        // 1. Remove any entries flagged for deletion by the compactor.
-        for id in result.toDelete {
-            if let entry = entry(withId: id) {
-                context.delete(entry)
-            }
-        }
-
-        // 2. `toUpdate` is a reference to an already-persisted entry that was
-        //    mutated in place by the compactor — nothing to do here besides
-        //    save, but we keep the branch explicit for clarity.
-        let persisted: MutationQueueEntry?
-        if let updated = result.toUpdate {
-            persisted = updated
-        } else if let toInsert = result.toInsert {
-            context.insert(toInsert)
-            persisted = toInsert
-        } else {
-            persisted = nil
-        }
-
-        save()
-        return persisted
     }
 
     // MARK: - Query
@@ -286,20 +223,6 @@ final class MutationQueue: MutationQueueProtocol {
         let inFlightRaw = MutationStatus.inFlight.rawValue
         var descriptor = FetchDescriptor<MutationQueueEntry>(
             predicate: #Predicate { $0.status == pendingRaw || $0.status == inFlightRaw }
-        )
-        descriptor.includePendingChanges = true
-        return (try? context.fetch(descriptor)) ?? []
-    }
-
-    private func pendingForEntity(entityType: String, entityId: String) -> [MutationQueueEntry] {
-        let pendingRaw = MutationStatus.pending.rawValue
-        var descriptor = FetchDescriptor<MutationQueueEntry>(
-            predicate: #Predicate {
-                $0.entityType == entityType
-                    && $0.entityId == entityId
-                    && $0.status == pendingRaw
-            },
-            sortBy: [SortDescriptor(\.createdAt, order: .forward)]
         )
         descriptor.includePendingChanges = true
         return (try? context.fetch(descriptor)) ?? []
